[UI] Moved network to settings window
