--- src/App/Network/CMakeLists.txt
@@ -1,3 +1,0 @@
-project(Network)
-
-add_library(Network NetworkWindow.cpp)
