--- src/App/Settings/Camera/CMakeLists.txt
@@ -1,3 +1,0 @@
-project(Camera)
-
-add_library(Camera CameraSettingsWidget.hpp CameraSettingsWidget.cpp)
