+++ src/App/Settings/CameraSettings/CMakeLists.txt
@@ -1,0 +1,3 @@
+project(CameraSettings)
+
+add_library(CameraSettings CameraSettingsWidget.hpp CameraSettingsWidget.cpp)
