+++ src/App/Settings/CameraSettings/CameraSettingsWidget.cpp
