+++ src/App/Settings/CameraSettings/CameraSettingsWidget.hpp
