--- src/App/Settings/Control/CMakeLists.txt
@@ -1,3 +1,0 @@
-project(Control)
-
-add_library(Control ControlSettingsWidget.hpp ControlSettingsWidget.cpp)
