+++ src/App/Settings/ControlSettings/CMakeLists.txt
@@ -1,0 +1,3 @@
+project(ControlSettings)
+
+add_library(ControlSettings ControlSettingsWidget.hpp ControlSettingsWidget.cpp)
