+++ src/App/Settings/MovementSettings/CMakeLists.txt
@@ -1,4 +1,4 @@
-project(Movement)
+project(MovementSettings)
 
 #do some cmake-protobuf magic
 find_package(Protobuf REQUIRED)
@@ -9,4 +9,4 @@ protobuf_generate_cpp(PROTO_SRCS PROTO_HDRS MovementSettingsMessage.proto)
 include_directories(${Protobuf_INCLUDE_DIRS})
 include_directories(${CMAKE_CURRENT_BINARY_DIR})
 
-add_library(Movement ${PROTO_SRCS} ${PROTO_HDRS} MovementSettingsWidget.hpp MovementSettingsWidget.cpp)
+add_library(MovementSettings ${PROTO_SRCS} ${PROTO_HDRS} MovementSettingsWidget.hpp MovementSettingsWidget.cpp)
