+++ src/App/Settings/NetworkSettings/CMakeLists.txt
@@ -1,0 +1,3 @@
+project(NetworkSettings)
+
+add_library(NetworkSettings NetworkSettingsWidget.hpp NetworkSettingsWidget.cpp)
